#129: Refactor encoding part of KafkaWriter into transformer component
